Encapsulating a component + adding svg shibasé
